// frontend/src/lib/sdk/models/ItemResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ItemResponse = {
    id: number;
    company_id: number;
    sku: string;
    name: string;
    type: string;
    hsn_sac: (string | null);
    uom: string;
    tax_rate: string;
    status: string;
    created_at: string;
    updated_at: string;
};