// frontend/src/lib/sdk/models/ItemUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ItemUpdate = {
    sku?: (string | null);
    name?: (string | null);
    type?: (string | null);
    hsn_sac?: (string | null);
    uom?: (string | null);
    tax_rate?: (number | string | null);
    status?: (string | null);
};